fix(web-monitor-server): AlarmScheduler，修复当category为0时，因未进行聚合判断导致计算错误的bug

+++ web-monitor-server/web-monitor-server-java/src/main/java/com/monitor/web/log/service/LogService.java
@@ -312,7 +312,7 @@ public class LogService extends ServiceBase {
                 }
                 if ("avg".equals(agg)) {
                     isExceedAlarmThreshold = uvRate / gap > val;
-                    actualValue = (float) uvRate / gap;
+                    actualValue = uvRate / gap;
                 }
             }
             if ("<".equals(op)) {
@@ -322,7 +322,7 @@ public class LogService extends ServiceBase {
                 }
                 if ("avg".equals(agg)) {
                     isExceedAlarmThreshold = uvRate / gap < val;
-                    actualValue = (float) uvRate / gap;
+                    actualValue = uvRate / gap;
                 }
             }
             // 环比昨天上涨、下跌
